Add new file in the modules folder
