+++ src/modules/todos.js
@@ -1,5 +1,5 @@
 export default class Todos {
-  static todo = "todos-todo-list-webpack-veronica365-2023";
+  static todo = 'todos-todo-list-webpack-veronica365-2023';
 
   static setTodo(todos) {
     const newTodos = JSON.stringify(todos);
@@ -17,7 +17,7 @@ export default class Todos {
     return todos || [];
   }
 
-  static removeData(id = "", allCompleted = false) {
+  static removeData(id = '', allCompleted = false) {
     let todos = this.listData();
     if (allCompleted) {
       todos = todos.filter((todo) => !todo.completed);
